+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.1] - 08-11-20
 ### Changed
 - no-std compatibility
+- export scalar as `BlsScalar`
 
 ## [0.1.5] - 29-10-20
 ### Changed